+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-07-19 03:16:21
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 00deb13b9d940fba4cf7b9d54c7db34bf97af684
+
+## [ForwardWidgets-2kuai] - 2025-07-19 03:16:21
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 973cabd15a0e3f08c3402312f38a531c88e65565
+
+## [ForwardWidgets-pack1r] - 2025-07-19 03:16:21
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 64224559c8760133446b8beeb338503413174ade
+
+
 ## [ForwardWidgets-huangxd] - 2025-07-18 03:22:18
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main